+++ src/components/common/Modal.tsx
@@ -1,8 +1,6 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { useEffect, useState } from 'react';
-import { createPortal } from 'react-dom';
-import { Z_INDEX } from 'constants/styles';
+import { ModalLayout } from 'components/layouts';
 
 interface ModalProps {
   isVisible: boolean;
@@ -19,65 +17,30 @@ export const Modal = ({
   onClose,
   onConfirm,
 }: ModalProps) => {
-  const [modalElement, setModalElement] = useState<HTMLElement | null>(null);
-
-  useEffect(() => {
-    const modalRoot = document.getElementById('modal-root');
-
-    setModalElement(modalRoot);
-  }, []);
-
-  if (modalElement === null) return null;
-
-  return createPortal(
-    isVisible ? (
-      <Layout>
-        <Overlay />
-        <Container>
-          <Message>{message}</Message>
-          <ButtonContainer>
-            <CancelButton type="button" onClick={onClose}>
-              취소
-            </CancelButton>
-            <ConfirmButton type="button" onClick={onConfirm}>
-              {confirmText}
-            </ConfirmButton>
-          </ButtonContainer>
-        </Container>
-      </Layout>
-    ) : null,
-    modalElement,
+  return (
+    <ModalLayout isVisible={isVisible}>
+      <Container>
+        <Message>{message}</Message>
+        <ButtonContainer>
+          <CancelButton type="button" onClick={onClose}>
+            취소
+          </CancelButton>
+          <ConfirmButton type="button" onClick={onConfirm}>
+            {confirmText}
+          </ConfirmButton>
+        </ButtonContainer>
+      </Container>
+    </ModalLayout>
   );
 };
 
-const Layout = styled.div`
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: ${Z_INDEX.modal};
-`;
-
-const Overlay = styled.div`
-  width: 100%;
-  height: 100%;
-  background-color: ${({ theme }) => theme.colors.black};
-  opacity: 60%;
-`;
-
 const Container = styled.div`
-  overflow: hidden;
   display: grid;
   grid-template-rows: auto 60px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
   width: 320px;
   border-radius: 16px;
   background-color: ${({ theme }) => theme.colors.white};
   ${({ theme }) => theme.fonts.body_03}
-  transform: translate(-50%, -50%);
 `;
 
 const Message = styled.p`
