+++ src/components/layouts/index.ts
@@ -1,3 +1,4 @@
 export { default as Layout } from './Layout';
 export { default as Navbar } from './Navbar';
 export * from './header';
+export * from './ModalLayout';
